refactor: split header into user main and promo parts
